Ditch ability to select individual disks from the Cinder proposal UI.
This code was fairly hackish and would not function properly in the
presence of the RAID barclamp making changes in any case.  For now,
leave us with the ability to select the first disk or all the disks,
along with the ability to fall back to a Large Flat File if need be.